requirements.txt
discord.py>=2.4.0
python-dotenv>=1.0.1
aiosqlite>=0.19.0
aiohttp>=3.9.5
uvloop>=0.19.0; platform_system != "Windows"
sentry-sdk>=2.8.0
pydantic>=2.7.4
pytest>=8.2.0
wavelink>=3.2.0

PyYAML>=6.0.1
